*: restructure package tree
This commit packages up individual modules and removes the top-level
main package.
