+++ .promu.yml
@@ -2,6 +2,9 @@ go: 1.6.2
 repository:
     path: github.com/prometheus/alertmanager
 build:
+    binaries:
+        - name: alertmanager
+          path: ./cmd/alertmanager
     flags: -a -tags netgo
     ldflags: |
         -X {{repoPath}}/vendor/github.com/prometheus/common/version.Version={{.Version}}
